+++ test/data/config/10-lifecycle-events-multiple/1.js
@@ -1,0 +1,4 @@
+module.exports = {
+	onClean: () => console.log('10-clean-1'),
+	onRestore: () => console.log('10-restore-1')
+};