+++ test/data/config/10-lifecycle-events-multiple/clean-package.config.js
@@ -1,0 +1,13 @@
+import { resolve } from 'path';
+
+
+module.exports = {
+	extends: [
+		resolve(process.cwd(), './1.js'),
+		resolve(process.cwd(), './2.js')
+	],
+	backupPath: './10-lifecycle-events-multiple.bak',
+	indent: '10',
+	remove: [],
+	replace: {}
+};