+++ test/data/config/9-lifecycle-events/clean-package.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+	backupPath: './9-lifecycle-events.bak',
+	indent: 9,
+	remove: [
+		'9a'
+	],
+	replace: {
+		'9b': 'abc'
+	},
+	onClean: () => '9-clean',
+	onRestore: () => '9-restore'
+};